+++ runtime/src/main/java/com/instana/android/view/InstanaComposableUiObserver.kt
@@ -9,8 +9,10 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.withFrameNanos
 import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import com.instana.android.Instana
@@ -24,17 +26,30 @@ object InstanaComposableUiObserver {
                 mutableStateOf(this.currentDestination?.route ?: "")
             }
 
+            val renderStartTime = remember { mutableStateOf(0L) }
+
             //Uses the composable's lifecycle
             DisposableEffect(this) {
                 val listener = NavController.OnDestinationChangedListener { _, destination, _ ->
                     currentScreen.value = destination.route ?: ""
-                    updateScreenName(currentScreen.value)
+                    renderStartTime.value = System.nanoTime() // Start measuring
                 }
                 addOnDestinationChangedListener(listener)
 
                 // Remove the listener when the composable is disposed to prevent leaks
                 onDispose {
                     removeOnDestinationChangedListener(listener)
+                }
+            }
+
+            // Second: Capture first frame render using withFrameNanos
+            LaunchedEffect(currentScreen.value) {
+                Instana.viewMeta.remove(ScreenAttributes.SCREEN_RENDERING_TIME.value)
+                withFrameNanos { frameTime ->
+                    val durationNanos = frameTime - renderStartTime.value
+                    val durationMillis = durationNanos / 1_000_000
+                    Instana.viewMeta.put(ScreenAttributes.SCREEN_RENDERING_TIME.value,durationMillis.toString())
+                    updateScreenName(currentScreen.value)
                 }
             }
     }